requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.10.0
av==16.1.0
bleach==6.3.0
blinker==1.9.0
cachetools==6.2.6
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cryptography==46.0.6
cycler==0.12.1
Cython==3.2.4
deprecation==2.1.0
dnspython==2.8.0
fastapi==0.117.1
filelock==3.20.0
Flask==3.0.0
fonttools==4.60.1
fsspec==2025.10.0
fvcore==0.1.5.post20221221
grad-cam==1.5.5
h11==0.16.0
h2==4.3.0
hf-xet==1.2.0
hpack==4.1.0
httpcore==1.0.9
httptools==0.6.4
httpx==0.28.1
huggingface_hub==1.4.1
hyperframe==6.1.0
idna==3.11
iopath==0.1.10
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.3
kaggle==2.0.0
kagglesdk==0.1.15
kiwisolver==1.4.9
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.6
mdurl==0.1.2
mmh3==5.2.1
mpmath==1.3.0
multidict==6.7.1
networkx==3.6
numpy==2.2.6
opencv-contrib-python==4.12.0.88
opencv-python==4.12.0.88
packaging==26.0
paho-mqtt==2.1.0
parameterized==0.9.0
pillow==11.3.0
polars==1.35.2
polars-runtime-32==1.35.2
portalocker==3.2.0
postgrest==2.28.3
propcache==0.4.1
protobuf==6.33.5
psutil==7.1.3
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
pyiceberg==0.11.1
PyJWT==2.12.1
pymongo==4.6.1
pyparsing==3.2.5
pyroaring==1.0.4
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-multipart==0.0.24
python-slugify==8.0.4
pytorchvideo==0.1.5
pywin32==311
PyYAML==6.0.2
realtime==2.28.3
reportlab==4.5.0
requests==2.32.5
rich==14.3.2
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.16.2
setuptools==80.9.0
shellingham==1.5.4
simplejson==3.20.2
six==1.17.0
sniffio==1.3.1
sounddevice==0.5.2
soundfile==0.13.1
starlette==0.48.0
storage3==2.28.3
StrEnum==0.4.15
strictyaml==1.7.3
supabase==2.28.3
supabase-auth==2.28.3
supabase-functions==2.28.3
sympy==1.14.0
tabulate==0.9.0
tenacity==9.1.4
termcolor==3.3.0
text-unidecode==1.3
threadpoolctl==3.6.0
timm==1.0.24
torch==2.10.0
torchvision==0.25.0
tqdm==4.67.3
ttach==0.0.3
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.1
typing_extensions==4.15.0
ultralytics==8.3.233
ultralytics-thop==2.0.18
urllib3==2.6.3
uvicorn==0.36.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
Werkzeug==3.1.7
yacs==0.1.8
yarl==1.23.0
zstandard==0.25.0
